fix(subscriptions/detail): update tests
