+++ frontend/tests/integration/components/project-info/component-test.js
